// src/api/models/OrganizationRepresentation.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OrganizationRepresentation = {
    id?: string;
    name?: string;
    displayName?: string;
    url?: string;
    realm?: string;
    domains?: Array<string>;
    attributes?: Record<string, Array<string>>;
};